requirements.txt
colorama==0.4.6
Pillow==9.3.0
qrcode==7.3.1
